2024/6/19

+++ backend/src/main/java/com/backend/service/board/BoardService.java
@@ -2,10 +2,12 @@ package com.backend.service.board;
 
 import com.backend.domain.board.Board;
 import com.backend.domain.board.BoardEditForm;
+import com.backend.domain.board.BoardImage;
 import com.backend.domain.board.BoardWriteForm;
-import com.backend.mapper.board.BoardFileMapper;
 import com.backend.mapper.board.BoardMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +19,9 @@ import software.amazon.awssdk.services.s3.model.ObjectCannedACL;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 @Service
@@ -25,18 +29,18 @@ import java.util.Objects;
 @RequiredArgsConstructor
 public class BoardService {
 
-    //@Value("${aws.s3.bucket.name}")
-    //String bucketName;
+    @Value("${aws.s3.bucket.name}")
+    String bucketName;
 
-    /*@Value("${image.src.prefix}")
-    String srcPrefix;*/
-
-    private final BoardFileMapper boardFileMapper;
-    private final BoardMapper mapper;
+    @Value("${image.src.prefix}")
+    String srcPrefix;
+    @Autowired
     private S3Client s3Client;
 
-    public void write(BoardWriteForm form, MultipartFile[] files, Authentication authentication
-    ) {
+    private final BoardMapper mapper;
+
+    public void write(BoardWriteForm form, Authentication authentication
+    ) throws IOException {
         Board board = new Board(
                 null,
                 Integer.valueOf(authentication.getName()),
@@ -44,84 +48,97 @@ public class BoardService {
                 form.getTitle(),
                 form.getContent(),
                 null
-
         );
         mapper.insert(board);
 
         Integer boardId = board.getId();
 
-        if (Objects.nonNull(boardId)) {
-            for (MultipartFile file : files) {
-                String fileName = file.getOriginalFilename();
-
-                try {
-                    // 파일 관련 테이블에 이미지 삽입
-                    boardFileMapper.insertImage(boardId, fileName);
-
-                    // S3에 이미지 저장
-                    saveBoardImageToS3(file, boardId, fileName);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
+        //사진
+        List<MultipartFile> images = form.getImages();
+        if (images != null) {
+            for (MultipartFile image : images) {
+                String imageName = image.getOriginalFilename();
+                mapper.insertImage(boardId, imageName);
+                saveBoardImageToS3(image, boardId, imageName);
             }
         }
     }
+
+    public Map<String, Object> findById(Integer boardId) {
+
+        Map<String, Object> result = new HashMap<>();
+
+        Board board = mapper.selectById(boardId);
+
+        if (board != null) {
+            return null;
+        }
+
+
+        List<String> imagesNames = mapper.selectImageNameById(boardId);
+        List<BoardImage> images = imagesNames.stream()
+                .map(imageName -> new BoardImage(imageName, STR."arbeit/board/\{boardId}/\{imageName}"))
+                .toList();
+
+        result.put("board", board);
+        result.put("images", images);
+        return result;
+
+
+    }
+
 
     public List<Board> list() {
         return mapper.selectAll();
     }
 
-    public Board findById(Integer id) {
-        return mapper.selectById(id);
-    }
+    public void edit(BoardEditForm form, Authentication authentication) throws IOException {
 
-    public void delete(Integer id) {
-
-        List<String> imageNames = boardFileMapper.selectImagesById(id);
-
-        for (String imageName : imageNames) {
-            removeBoardImageToS3(id, imageName);
-        }
-
-        boardFileMapper.deleteByImagesId(id);
-        mapper.deleteById(id);
-    }
-
-    public void edit(BoardEditForm form, Integer id, List<String> removeImages, MultipartFile[] addImages) throws IOException {
+        Integer boardId = form.getId();
         Board board = new Board(
-                id,
+                boardId,
                 form.getTitle(),
                 form.getContent()
         );
         mapper.update(board);
 
-        if (Objects.nonNull(removeImages) && !removeImages.isEmpty()) {
-            for (String imageName : removeImages) {
-
-                removeBoardImageToS3(id, imageName);
-                boardFileMapper.deleteImageByIdAndName(id, imageName);
+        //사진삭제
+        removeBoardImageToS3(boardId, form.getRemoveImages());
+        List<String> removeImages = form.getRemoveImages();
+        if (removeImages != null) {
+            for (String removeImage : removeImages) {
+                mapper.deleteByboardIdAndImageName(boardId, removeImage);
             }
         }
-        if (Objects.nonNull(addImages) && addImages.length > 0) {
-            for (MultipartFile file : addImages) {
-                String fileName = file.getOriginalFilename();
-                boardFileMapper.insertImage(id, fileName);
-                saveBoardImageToS3(file, id, fileName);
+        //사진 추가
+        List<MultipartFile> addImages = form.getAddImages();
+        if (addImages != null) {
+            for (MultipartFile addImage : addImages) {
+                String imageName = addImage.getOriginalFilename();
+                mapper.insertImage(boardId, imageName);
+                saveBoardImageToS3(addImage, boardId, addImage.getOriginalFilename());
             }
         }
+
     }
 
-    public boolean hasAccess(Integer id, Authentication authentication) {
-        Integer memberId = mapper.selectById(id).getMemberId();
-        Integer loginId = Integer.valueOf(authentication.getName());
-        return Objects.equals(memberId, loginId);
+
+    public void delete(Integer boardId) {
+        //파일명
+        List<String> fileNames = mapper.selectImageNameById(boardId);
+        removeBoardImageToS3(boardId, fileNames);
+        //사진
+        mapper.deleteByboardId(boardId);
+        //글
+        mapper.deleteById(boardId);
     }
 
-    private void saveBoardImageToS3(MultipartFile image, int boardId, String filename) throws IOException {
-        //String key = STR."arbeit/board/\{boardId}/\{filename}";
+
+    private void saveBoardImageToS3(MultipartFile image, int boardId, String imageName) throws IOException {
+        String key = STR."arbeit/board/\{boardId}/\{imageName}";
         PutObjectRequest objectRequest = PutObjectRequest.builder()
-                //.bucket(bucketName)
-                //.key(key)
+                .bucket(bucketName)
+                .key(key)
                 .acl(ObjectCannedACL.PUBLIC_READ)
                 .build();
 
@@ -129,12 +146,27 @@ public class BoardService {
                 RequestBody.fromInputStream(image.getInputStream(), image.getSize()));
     }
 
-    private void removeBoardImageToS3(Integer boardId, String imageName) {
-        //String key = STR."arbeit/board/\{boardId}/\{imageName}";
-        DeleteObjectRequest objectRequest = DeleteObjectRequest.builder()
-                //.bucket(bucketName)
-                //.key(key)
-                .build();
-        s3Client.deleteObject(objectRequest);
+    private void removeBoardImageToS3(Integer boardId, List<String> removeImages) {
+        if (removeImages != null && !removeImages.isEmpty()) {
+            for (String imageName : removeImages) {
+
+                String key = STR."arbeit/board/\{boardId}/\{imageName}";
+                DeleteObjectRequest objectRequest = DeleteObjectRequest.builder()
+                        .bucket(bucketName)
+                        .key(key)
+                        .build();
+                s3Client.deleteObject(objectRequest);
+
+                //db.삭제
+                mapper.deleteByboardIdAndImageName(boardId, imageName);
+            }
+        }
+    }
+
+
+    public boolean hasAccess(BoardEditForm form, Authentication authentication) {
+        Integer memberId = mapper.selectById(form.getMemberId()).getMemberId();
+        Integer loginId = Integer.valueOf(authentication.getName());
+        return Objects.equals(memberId, loginId);
     }
 }
